+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "gerocuidado",
   "version": "1.0.0",
-  "main": "expo-router/entry",
+  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
